Smaller Count And Threshold (#117)
As previously discussed, and to match the official FROST implementation
from the ZCash Foundation, `count` and `threshold` are now `uint16`s.
This unfortunately added quite a bit of small changes as `viem` maps
`uint64`s to `bigint`s but `uint16`s to `number`s.

There are a couple of other small changes that I did while I was at it,
in particular I made all the `KeyGen` events also include a "completed"
flag for each phase ("committed", "shared", "compromised", and
"confirmed"), and improved the typing in `logToTransition` to prevent
issues where we forget to add an event to the ABI (would be a compiler
error in that case).

Futhermore, I fixed a small issue with the event listener which was
causing complaint flow events to not be processed (as they were missing
from the contract ABI).

+++ contracts/src/libraries/FROSTGroupId.sol
@@ -33,7 +33,7 @@ library FROSTGroupId {
      * @param context The context data for the group.
      * @return result The computed group ID.
      */
-    function create(bytes32 participants, uint64 count, uint64 threshold, bytes32 context)
+    function create(bytes32 participants, uint16 count, uint16 threshold, bytes32 context)
         internal
         pure
         returns (T result)
